# docs/openapi.json
{
  "components" : {
    "schemas" : {
      "ActivityResponse" : {
        "properties" : {
          "activityDate" : {
            "example" : "2025-06-10",
            "format" : "date",
            "type" : "string"
          },
          "cost" : {
            "example" : 1.25E7,
            "type" : "number"
          },
          "id" : {
            "example" : 31,
            "format" : "int64",
            "type" : "integer"
          },
          "note" : {
            "description" : "Có thể null, trừ hoạt động loại OTHER",
            "example" : "Bón NPK 16-16-8, 20 bao",
            "type" : "string"
          },
          "seasonId" : {
            "description" : "Niên vụ hệ thống tự gán theo ngày thực hiện",
            "example" : 7,
            "format" : "int64",
            "type" : "integer"
          },
          "seasonLabel" : {
            "example" : "2025/2026",
            "type" : "string"
          },
          "type" : {
            "enum" : [ "WATERING", "FERTILIZING", "SPRAYING", "WEEDING", "PRUNING", "OTHER" ],
            "example" : "FERTILIZING",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "HarvestResponse" : {
        "properties" : {
          "harvestDate" : {
            "example" : "2025-11-28",
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "example" : 12,
            "format" : "int64",
            "type" : "integer"
          },
          "pricePerKg" : {
            "description" : "Doanh thu chia sản lượng — giá bán thực tế của lần thu hoạch này",
            "example" : 24000.0,
            "type" : "number"
          },
          "quantityKg" : {
            "example" : 3200.5,
            "format" : "double",
            "type" : "number"
          },
          "revenue" : {
            "example" : 7.68E7,
            "type" : "number"
          },
          "seasonId" : {
            "description" : "Niên vụ hệ thống tự gán theo ngày thu hoạch",
            "example" : 7,
            "format" : "int64",
            "type" : "integer"
          },
          "seasonLabel" : {
            "example" : "2025/2026",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "LogActivityRequest" : {
        "properties" : {
          "activityDate" : {
            "description" : "Ngày thực hiện; quyết định niên vụ của bản ghi",
            "example" : "2025-06-10",
            "format" : "date",
            "type" : "string"
          },
          "cost" : {
            "description" : "Chi phí (VND). Bỏ trống = 0 khi tự làm, không tốn chi phí.",
            "example" : 12500000,
            "minimum" : 0,
            "type" : "number"
          },
          "note" : {
            "description" : "Ghi chú; bắt buộc khi loại hoạt động là OTHER",
            "example" : "Bón NPK 16-16-8, 20 bao",
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "type" : {
            "description" : "WATERING: tưới · FERTILIZING: bón phân · SPRAYING: phun thuốc · WEEDING: làm cỏ · PRUNING: tỉa cành · OTHER: khác (bắt buộc ghi chú)",
            "enum" : [ "WATERING", "FERTILIZING", "SPRAYING", "WEEDING", "PRUNING", "OTHER" ],
            "example" : "FERTILIZING",
            "type" : "string"
          }
        },
        "required" : [ "activityDate", "type" ],
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RecordHarvestRequest" : {
        "properties" : {
          "harvestDate" : {
            "description" : "Ngày thu hoạch; quyết định niên vụ của bản ghi",
            "example" : "2025-11-28",
            "format" : "date",
            "type" : "string"
          },
          "quantityKg" : {
            "description" : "Sản lượng tươi hoặc khô tùy cách nhà nông ghi, đơn vị kg",
            "example" : 3200.5,
            "exclusiveMinimum" : 0,
            "format" : "double",
            "maximum" : 100000000,
            "type" : "number"
          },
          "revenue" : {
            "description" : "Doanh thu (VND). Bỏ trống = 0 khi đã thu hoạch nhưng chưa bán.",
            "example" : 76800000,
            "minimum" : 0,
            "type" : "number"
          }
        },
        "required" : [ "harvestDate", "quantityKg" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Quản lý canh tác nông trại: nông trại, lô đất, danh mục cây trồng, lứa trồng xen canh, nhật ký chăm sóc, thu hoạch và lãi/lỗ theo từng cây.\n\nLỗi trả về dạng RFC 9457 ProblemDetail; trường `rule` là mã quy tắc nghiệp vụ trong docs/architecture.md.",
    "title" : "Farm Management System API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/activities/{activityId}" : {
      "delete" : {
        "description" : "Nhật ký được xóa thật vì đây là dữ liệu nhập tay, không phải sự kiện vòng đời.",
        "operationId" : "delete_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "activityId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Xóa hoạt động nhập sai",
        "tags" : [ "6. Nhật ký canh tác" ]
      },
      "get" : {
        "operationId" : "get_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "activityId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActivityResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Chi tiết một hoạt động",
        "tags" : [ "6. Nhật ký canh tác" ]
      },
      "put" : {
        "description" : "Đổi ngày sang chu kỳ sản xuất khác thì bản ghi tự chuyển sang niên vụ tương ứng.",
        "operationId" : "correct_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "activityId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LogActivityRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActivityResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Sửa hoạt động nhập sai",
        "tags" : [ "6. Nhật ký canh tác" ]
      }
    },
...
    "/api/v1/harvests/{harvestId}" : {
      "delete" : {
        "operationId" : "delete_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "harvestId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Xóa lần thu hoạch nhập sai",
        "tags" : [ "7. Thu hoạch" ]
      },
      "get" : {
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "harvestId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HarvestResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Chi tiết một lần thu hoạch",
        "tags" : [ "7. Thu hoạch" ]
      },
      "put" : {
        "description" : "Không đảo ngược trạng thái lứa trồng về GROWING (BR-09).",
        "operationId" : "correct_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "harvestId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecordHarvestRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HarvestResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Sửa lần thu hoạch nhập sai",
        "tags" : [ "7. Thu hoạch" ]
      }
    },
    "/api/v1/plantings" : {
      "get" : {
        "description" : "Cho form ghi nhật ký chọn lứa trồng trong một lần gọi. farmId bỏ trống: mọi nông trại. Nông trại của người khác trả 404 (BR-11). Xếp theo tên lô, lứa mới trồng trước.",
        "operationId" : "listOwned",
        "parameters" : [ {
          "in" : "query",
          "name" : "farmId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "activeOnly",
          "required" : false,
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PlantingResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Mọi lứa trồng của tôi",
        "tags" : [ "4. Lứa trồng" ]
      }
    },
...
    "/api/v1/plantings/{plantingId}/production-start" : {
      "post" : {
        "description" : "GROWING → PRODUCING (BR-03).",
        "operationId" : "startProducing",
        "parameters" : [ {
          "in" : "path",
          "name" : "plantingId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlantingResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Bắt đầu cho thu hoạch",
        "tags" : [ "4. Lứa trồng" ]
      }
    },
...
    "/api/v1/plantings/{plantingId}/termination" : {
      "post" : {
        "description" : "Bắt buộc ngày và lý do; ngày kết thúc ≥ ngày trồng và ≤ hôm nay (BR-04). Lứa trồng đã kết thúc không chuyển trạng thái được nữa (BR-03).",
        "operationId" : "terminate",
        "parameters" : [ {
          "in" : "path",
          "name" : "plantingId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Id chủ nông trại (bỏ trống = người dùng mặc định 1). Thay bằng JWT ở Phase 5.",
          "in" : "header",
          "name" : "X-User-Id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TerminatePlantingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlantingResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Dữ liệu không hợp lệ — lỗi từng trường nằm trong 'errors', lỗi tham số nằm trong 'detail'"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Kết thúc lứa trồng (cưa bỏ)",
        "tags" : [ "4. Lứa trồng" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}